@@ -2,7 +2,7 @@
   "name": "magicMirror",
   "version": "1.0.0",
   "scripts": {
-    "build-ui": "webpack services/ui/js/index.jsx services/ui/public/bundle.js",
+    "build-ui": "webpack",
     "mocha": "mocha",
     "start": "node index.js"
   },
@@ -17,11 +17,14 @@
     "babel-preset-react": "^6.11.1",
     "bluebird": "^3.4.1",
     "classnames": "*",
-    "eslint": "^3.1.1",
+    "css-loader": "^0.28.4",
+    "eslint": "^3.19.0",
     "eslint-plugin-lodash": "^1.10.0",
     "eslint-plugin-react": "^5.2.2",
     "event-bus": "git+https://bitbucket.org/karleberts/eventbus.git",
     "express": "^4.14.0",
+    "extract-text-webpack-plugin": "^2.1.2",
+    "file-loader": "^0.11.2",
     "google-client-api": "*",
     "googleapis": "^12.0.0",
     "handlebars": "*",
@@ -32,7 +35,7 @@
     "mocha": "*",
     "moment": "^2.14.1",
     "node-fetch": "^1.6.3",
-    "node-sass": "^3.8.0",
+    "node-sass": "^3.13.1",
     "ramda": "^0.21.0",
     "react": "^15.2.1",
     "react-addons-css-transition-group": "^15.2.1",
@@ -48,6 +51,8 @@
     "rx": "*",
     "rx-react": "*",
     "rxjs": "^5.0.0-beta.10",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.18.2",
     "webpack": "*",
     "webpack-dev-server": "^1.14.1",
     "ws": "^1.1.1"
