@@ -1,7 +1,7 @@
 'use strict';
 const Rx = require('rxjs');
 
-const WebSocket = (typeof window !== 'undefined') ? window.WebSocket : require('ws');
+const WebSocket = (typeof window !== 'undefined') ? window.WebSocket : require('uws');
 
 /// by default, when a message is received from the server, we are trying to decode it as JSON
 /// we can override it in the constructor
