@@ -28,6 +28,7 @@ CAMERA_INST = None
 RAW_CAPTURE = None
 CAPTURE_INTERVAL = 350
 FRAME_COUNT_UNTIL_FALSE = 15 # num. of frames to see 0 faces in until emitting a 'not found' result (~5s?)
+SHOW_CV = False;
 
 timer = Observable.interval(CAPTURE_INTERVAL)
 
@@ -44,6 +45,7 @@ def get_image_as_array(camera):
     return image
 def find_faces_in_image(image):
     """Find faces in the given opencv numpy array"""
+    global SHOW_CV
     # t = time.time()
     gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
     # u = time.time()
@@ -53,12 +55,13 @@ def find_faces_in_image(image):
     faces = FACE_CASCADE.detectMultiScale(
         gray,
         scaleFactor=1.1,
-        minNeighbors=2,
+        minNeighbors=4,
         minSize=(30, 30)
     )
-    # for (x, y, w, h) in faces:
-        # cv2.rectangle(gray, (x, y), (x+w, y+h), (255, 0, 0), 2)
-    # cv2.imshow("image", gray)
+    if SHOW_CV:
+        for (x, y, w, h) in faces:
+            cv2.rectangle(gray, (x, y), (x+w, y+h), (255, 0, 0), 2)
+        cv2.imshow("image", gray)
     # key = cv2.waitKey(1) & 0xFF
     # u = time.time()
     # elapsed = u - t
@@ -67,17 +70,21 @@ def find_faces_in_image(image):
 def do_face_check():
     """event_bus message handler"""
     global RAW_CAPTURE
+    global SHOW_CV
+    # print('doing a face check')
     camera = get_camera()
     try:
         image = get_image_as_array(camera)
         face_count = find_faces_in_image(image)
         RAW_CAPTURE.truncate(0)
-        camera.annotate_text = ('analog_gain: {}, brightness: {}, contrast: {}, exposure_speed: {}, faces: {}'
-                                .format(float(camera.analog_gain), float(camera.brightness), camera.brightness, camera.exposure_speed, face_count))
+        if SHOW_CV:
+            camera.annotate_text = ('analog_gain: {}, brightness: {}, contrast: {}, exposure_speed: {}, faces: {}'
+                                    .format(float(camera.analog_gain), float(camera.brightness), camera.brightness, camera.exposure_speed, face_count))
     except (KeyboardInterrupt, SystemExit):
         raise
     except Exception:
         camera.close()
+        ioloop.IOLoop.current(True).stop()
         sys.exit()
     # print("found {0} faces".format(face_count))
     return face_count > 0
@@ -89,12 +96,11 @@ def get_camera(should_create=True):
     if CAMERA_INST is not None:
         return CAMERA_INST
     elif should_create:
-        # cv2.namedWindow('image', cv2.WINDOW_NORMAL)
         CAMERA_INST = PiCamera(resolution='480x368')
         CAMERA_INST.exposure_mode = 'night'
         RAW_CAPTURE = PiRGBArray(CAMERA_INST, size=(480, 368))
         print('camera created')
-        # CAMERA_INST.start_preview()
+        #CAMERA_INST.start_preview()
         return CAMERA_INST
     else:
         return CAMERA_INST
@@ -102,9 +108,11 @@ def get_camera(should_create=True):
 def close_camera():
     """if a camera is instantiated, destroy it"""
     global CAMERA_INST
+    global SHOW_CV
     if CAMERA_INST is not None:
         print('closing the camera')
-        # cv2.destroyWindow('image')
+        if SHOW_CV:
+            cv2.destroyWindow('image')
         # CAMERA_INST.stop_preview()
         CAMERA_INST.close()
         CAMERA_INST = None
@@ -119,7 +127,21 @@ def stop():
         print('camera closed')
     else:
         print('camera already closed')
+    ioloop.IOLoop.current(True).stop()
     sys.exit()
+
+def handle_debug_req(req):
+    global SHOW_CV
+    if (SHOW_CV is not req['params']):
+        SHOW_CV = req['params']
+        if (SHOW_CV):
+            print('opening cv window {}'.format(SHOW_CV))
+            cv2.namedWindow('image', cv2.WINDOW_NORMAL)
+            CAMERA_INST.start_preview()
+        else:
+            print('closing cv window {}'.format(SHOW_CV))
+            cv2.destroyWindow('image')
+            CAMERA_INST.stop_preview()
 
 # def test(msg):
 #     print('testing a message')
@@ -148,10 +170,14 @@ def start():
     signal.signal(signal.SIGTERM, lambda sig, frame: stop())
     pause_stream = listen_for_pause()
     time.sleep(1)
+    last_result = False;
+    def handle_result(res):
+        last_result = res;
+        EVENT_BUS.send_message('faceDetect.result', res)
     result_stream = (timer
                      .with_latest_from(pause_stream, lambda x, is_paused: is_paused)
                      .filter(lambda is_paused: is_paused is not True)
-                     .map(lambda is_paused: do_face_check())
+                     .map(lambda x: do_face_check())
                      .share())
     found_stream = (result_stream
                     .filter(lambda x: x is True))
@@ -160,8 +186,18 @@ def start():
                         .map(lambda res: False))
     (Observable.merge(found_stream, not_found_stream)
      .distinct_until_changed()
-     .subscribe(lambda res: EVENT_BUS.send_message('faceDetect.result', res)))
+     .subscribe(handle_result))
     pause_stream.connect()
+
+    #show/hide the debug window
+    (EVENT_BUS.request_stream
+     .filter(lambda req: req['topic'] == 'faceDetect.showDebug')
+     .subscribe(handle_debug_req))
+
+    (EVENT_BUS.request_stream
+     .filter(lambda req: req['topic'] == 'faceDetect.getStatus')
+     .subscribe(lambda req: req['respond'](last_result)))
+
     EVENT_BUS.send_message('faceDetect.ready', True, 'magicMirror')
 
 
@@ -170,7 +206,6 @@ def main():
     EVENT_BUS.connect(ENDPOINT_ID, start)
     print('running')
 
-
 if __name__ == '__main__':
     main()
-    ioloop.IOLoop.instance().start()
+    ioloop.IOLoop.current(True).start()
