@@ -1,18 +1,8 @@
 'use strict';
-const { execSync } = require('child_process');
 const { requests } = require('event-bus/client');
 
-function sleepHdmi () {
-	execSync('tvservice -o', err => {
-		if (err) { throw(err); }
-	});
-}
+const { sleepHdmi, wakeHdmi } = require('../../lib/hdmi');
 
-function wakeHdmi () {
-	execSync('tvservice -o', err => {
-		if (err) { throw(err); }
-	});
-}
 
 requests
 	.filter(req => req.topic === 'hdmi.sleep' && req.params)
