@@ -4,6 +4,7 @@ const React = require('react');
 const { render } = require('react-dom');
 const eventBus = window.eb = require('event-bus/client');
 
+import '../scss/main.scss';
 const Mirror = require('./components/mainPage.jsx');
 const configureStore = require('./redux/configureStore');
 //TODO- global ref
@@ -14,8 +15,11 @@ if (window.devToolsExtension) {
 	window.devToolsExtension.updateStore(store);
 }
 
-eventBus.connect('magicMirror.ui')
-	.then(() => console.log('connected to evtBus'));
+eventBus.connect('magicMirror.ui', true)
+	.then(() => {
+		eventBus.sendMessage('ui.ready', true, 'magicMirror');
+		console.log('connected to evtBus');
+	});
 
 
 var container = document.getElementById('content');
